+++ app/src/rec/dining.py
@@ -1,0 +1,7 @@
+from flask import Blueprint, render_template
+
+dining_bp = Blueprint('dining', __name__)
+
+@dining_bp.route('/dining')
+def dining():
+    return render_template('dining.html')
