+++ app/templates/home.html
@@ -12,7 +12,7 @@
                 <div class="card-body">
                     <h5 class="card-title">Movies</h5>
                     <p class="card-text">Find the latest and greatest movies to watch!</p>
-                    <a href="{{ url_for('movies') }}" class="btn btn-primary">Explore Movies</a>
+                    <a href="{{ url_for('movies.movies') }}" class="btn btn-primary">Explore Movies</a>
                 </div>
             </div>
         </div>
@@ -22,7 +22,7 @@
                 <div class="card-body">
                     <h5 class="card-title">Dining</h5>
                     <p class="card-text">Check out the top dining spots in your area.</p>
-                    <a href="{{ url_for('dining') }}" class="btn btn-primary">Explore Dining</a>
+                    <a href="{{ url_for('dining.dining') }}" class="btn btn-primary">Explore Dining</a>
                 </div>
             </div>
         </div>
@@ -32,7 +32,7 @@
                 <div class="card-body">
                     <h5 class="card-title">Activities</h5>
                     <p class="card-text">Discover exciting activities and adventures!</p>
-                    <a href="{{ url_for('activities') }}" class="btn btn-primary">Explore Activities</a>
+                    <a href="{{ url_for('activities.activities') }}" class="btn btn-primary">Explore Activities</a>
                 </div>
             </div>
         </div>
